Move the EventSchedulerValidator to the eventscheduler package
Move the EventSchedulerValidator to the eventscheduler package

#691
